CLAN6-158 Rename config packages
