+++ cogboard-app/src/main/kotlin/com/cognifide/cogboard/config/ConfigLoader.kt
@@ -1,14 +1,14 @@
 package com.cognifide.cogboard.config
 
 import com.cognifide.cogboard.CogboardConstants
-import com.cognifide.cogboard.config.strategy.Config
-import com.cognifide.cogboard.config.strategy.ConfigFactory
+import com.cognifide.cogboard.config.type.Config
+import com.cognifide.cogboard.config.type.ConfigFactory
 import io.vertx.core.json.JsonObject
 import java.io.File
 
 class ConfigLoader(private val configType: ConfigType){
 
-    private var config: Config = ConfigFactory.resolveStrategy(configType)
+    private var config: Config = ConfigFactory.resolveByType(configType)
 
     fun loadConfig(): JsonObject {
         val conf = File(config.filePath()).readText()
