+++ cogboard-app/src/main/kotlin/com/cognifide/cogboard/config/type/ConfigFactory.kt
@@ -1,11 +1,11 @@
-package com.cognifide.cogboard.config.strategy
+package com.cognifide.cogboard.config.type
 
 import com.cognifide.cogboard.config.ConfigType
 import com.cognifide.cogboard.config.ConfigType.*
 
 class ConfigFactory {
     companion object {
-        fun resolveStrategy(configType: ConfigType): Config {
+        fun resolveByType(configType: ConfigType): Config {
             return when (configType) {
                 BOARDS -> BoardsConfig()
                 ENDPOINTS -> EndpointsConfig()
